refactor: 일자리 지원 정보 관련 API 분리
