+++ src/main/java/com/becareful/becarefulserver/domain/caregiver/controller/CaregiverController.java
@@ -29,7 +29,6 @@ import java.util.List;
 public class CaregiverController {
 
     private final CaregiverService caregiverService;
-    private final WorkApplicationService workApplicationService;
     private final CareerService careerService;
     private final CompletedMatchingService completedMatchingService;
 
@@ -56,34 +55,6 @@ public class CaregiverController {
         return ResponseEntity.ok(response);
     }
 
-    @Operation(summary = "일자리 신청 정보 등록/수정")
-    @PutMapping("/work-application")
-    public ResponseEntity<Void> updateWorkApplication(@Valid @RequestBody WorkApplicationUpdateRequest request) {
-        workApplicationService.updateWorkApplication(request);
-        return ResponseEntity.ok().build();
-    }
-
-    @Operation(summary = "일자리 신청 정보 조회", description = "신청 정보를 등록한 적이 없다면 빈 값을 반환합니다.")
-    @GetMapping("/work-application")
-    public ResponseEntity<WorkApplicationResponse> getWorkApplication() {
-        WorkApplicationResponse response = workApplicationService.getWorkApplication();
-        return ResponseEntity.ok(response);
-    }
-
-    @Operation(summary = "일자리 신청 활성화", description = "등록한 일자리 신청 내용을 기반으로 매칭을 활성화 합니다.")
-    @PostMapping("/work-applicatioin/active")
-    public ResponseEntity<Void> updateWorkApplicationActive() {
-        workApplicationService.updateWorkApplicationActive();
-        return ResponseEntity.ok().build();
-    }
-
-    @Operation(summary = "일자리 신청 비활성화", description = "일자리 매칭을 비활성화 합니다.")
-    @PostMapping("/work-applicatioin/inactive")
-    public ResponseEntity<Void> updateWorkApplicationInactive() {
-        workApplicationService.updateWorkApplicationInactive();
-        return ResponseEntity.ok().build();
-    }
-
     @Operation(summary = "경력서 등록/수정")
     @PutMapping("/career")
     public ResponseEntity<Void> updateCareer(@Valid @RequestBody CareerUpdateRequest request) {
@@ -105,22 +76,19 @@ public class CaregiverController {
         return ResponseEntity.ok(response);
     }
 
-    //TODO
     @Operation(summary = "확정된 일자리의 리스트가 반환됩니다.")
     @GetMapping("/my/completed-matching-list")
     public ResponseEntity<List<CompletedMatchingInfoResponse>> getCompletedMatchingsByCaregiverId() {
             List<CompletedMatchingInfoResponse> responseList = completedMatchingService.getCompletedMatchings();
             return ResponseEntity.ok(responseList);
-        }
+    }
 
-    //TODO
     @Operation(summary = "나의 일자리 화면에서 메모 수정")
     @PutMapping("/my/complete-matching-list/{completedMatchingId}")
     public ResponseEntity<Void> editCompletedMatchingMemo(@PathVariable Long completedMatchingId, @RequestBody EditCompletedMatchingNoteRequest request){
         completedMatchingService.editNote(completedMatchingId, request);
         return ResponseEntity.ok().build();
     }
-
 
     @Operation(summary = "요양보호사 채팅 목록")
     @GetMapping("/chat/list")
